Extract EventHandlerInterface to separate file
Moved interface definition from BaseEventHandler.php to its own file.
Renamed EventHandler -> EventHandlerInterface for clarity.
